refactor struct at rule
