+++ src/rules/pylint/rules/mod.rs
@@ -5,7 +5,7 @@ pub use invalid_all_object::{invalid_all_object, InvalidAllObject};
 pub use magic_value_comparison::magic_value_comparison;
 pub use merge_isinstance::merge_isinstance;
 pub use property_with_parameters::property_with_parameters;
-pub use too_many_args::too_many_args;
+pub use too_many_args::{too_many_args, TooManyArgs};
 pub use unnecessary_direct_lambda_call::unnecessary_direct_lambda_call;
 pub use use_from_import::use_from_import;
 pub use use_sys_exit::use_sys_exit;
